refactor: delete filing/balance_due.py (folded into forms/f4868.py)
Swap orchestrator import from tenforty.filing.balance_due to
tenforty.forms.f4868, update tests/invariants.py likewise, and
git-rm the legacy module and its test file.
